refactor(video-list-item): hookify and reorganize component
