+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/container.jsx
@@ -34,7 +34,11 @@ export default withTracker((props) => {
     voiceUser: VoiceUsers.findOne({ intId: userId },
       { fields: { muted: 1, listenOnly: 1, talking: 1 } }),
     user: Users.findOne({ intId: userId },
-      { fields: { pin: 1, userId: 1 } }),
+      {
+        fields: {
+          pin: 1, userId: 1, name: 1,
+        },
+      }),
   };
 })(VideoListItemContainer);
