+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/styles.js
@@ -1,21 +1,10 @@
 import styled from 'styled-components';
-import Icon from '/imports/ui/components/common/icon/component';
-import Button from '/imports/ui/components/common/button/component';
 import {
   colorPrimary,
   colorBlack,
   colorWhite,
-  colorOffWhite,
-  colorDanger,
-  colorSuccess,
-  colorTransparent,
 } from '/imports/ui/stylesheets/styled-components/palette';
-import { TextElipsis, DivElipsis } from '/imports/ui/stylesheets/styled-components/placeholders';
-import { landscape, mediumUp } from '/imports/ui/stylesheets/styled-components/breakpoints';
-import {
-  audioIndicatorWidth,
-  audioIndicatorFs,
-} from '/imports/ui/stylesheets/styled-components/general';
+import { TextElipsis } from '/imports/ui/stylesheets/styled-components/placeholders';
 
 const Content = styled.div`
   position: relative;
@@ -161,149 +150,24 @@ const Video = styled.video`
   `}
 `;
 
-const Info = styled.div`
+const TopBar = styled.div`
   position: absolute;
-  bottom: 1px;
-  left: 7px;
-  right: 5px;
-  z-index: 2;
-`;
-
-const Dropdown = styled.div`
   display: flex;
-  outline: none !important;
-  width: 70%;
-
-  @media ${mediumUp} {
-    >[aria-expanded] {
-      padding: .25rem;
-    }
-  }
-
-  @media ${landscape} {
-    button {
-      width: calc(100vw - 4rem);
-      margin-left: 1rem;
-    }
-  }
-
-  ${({ isFirefox }) => isFirefox && `
-    max-width: 100%;
-  `}
-`;
-
-const UserName = styled(TextElipsis)`
-  position: relative;
-  max-width: 75%;
-  // Keep the background with 0.5 opacity, but leave the text with 1
-  background-color: rgba(0, 0, 0, 0.5);
-  border-radius: 1px;
-  color: ${colorOffWhite};
-  padding: 0 1rem 0 .5rem !important;
-  font-size: 80%;
-
-  ${({ noMenu }) => noMenu && `
-    padding: 0 .5rem 0 .5rem !important;
-  `}
-`;
-
-const Muted = styled(Icon)`
-  display: inline-block;
-  position: absolute;
-  right: 7px;
-  bottom: 6px;
-  width: ${audioIndicatorWidth};
-  height: ${audioIndicatorWidth};
-  min-width: ${audioIndicatorWidth};
-  min-height: ${audioIndicatorWidth};
-  color: ${colorWhite};
-  border-radius: 50%;
-
-  &::before {
-    font-size: ${audioIndicatorFs};
-  }
-
-  background-color: ${colorDanger};
-`;
-
-const Voice = styled(Icon)`
-  display: inline-block;
-  position: absolute;
-  right: 7px;
-  bottom: 6px;
-  width: ${audioIndicatorWidth};
-  height: ${audioIndicatorWidth};
-  min-width: ${audioIndicatorWidth};
-  min-height: ${audioIndicatorWidth};
-  color: ${colorWhite};
-  border-radius: 50%;
-
-  &::before {
-    font-size: ${audioIndicatorFs};
-  }
-
-  background-color: ${colorSuccess};
-`;
-
-const DropdownTrigger = styled(DivElipsis)`
-  position: relative;
-  max-width: 75%;
-  // Keep the background with 0.5 opacity, but leave the text with 1
-  background-color: rgba(0, 0, 0, 0.5);
-  border-radius: 1px;
-  color: ${colorOffWhite};
-  padding: 0 1rem 0 .5rem !important;
-  font-size: 80%;
-
-  cursor: pointer;
-
-  &::after {
-    content: "\\203a";
-    position: absolute;
-    transform: rotate(90deg);
-    top: 45%;
-    width: 0;
-    line-height: 0;
-    right: .45rem;
-  }
-`;
-
-const PinButtonWrapper = styled.div`
-  position: absolute;
+  width: 100%;
+  z-index: 1;
   top: 0;
-  left: auto;
-  background-color: rgba(0,0,0,.3);
-  cursor: pointer;
-  border: 0;
-  z-index: 2;
-  margin: 2px;
-
-  [dir="rtl"] & {
-    right: auto;
-    left :0;
-  }
-
-  [class*="presentationZoomControls"] & {
-    position: relative !important;
-  }
+  padding: 5px;
+  justify-content: space-between;
 `;
 
-const PinButton = styled(Button)`
-  padding: 5px;
-  &,
-  &:active,
-  &:hover,
-  &:focus {
-    background-color: ${colorTransparent} !important;
-    border: none !important;
-
-    & > i {
-      border: none !important;
-      color: ${colorWhite};
-      font-size: 1rem;
-      background-color: ${colorTransparent} !important;
-    }
-  }
+const BottomBar = styled.div`
+  position: absolute;
+  display: flex;
+  width: 100%;
+  z-index: 1;
+  bottom: 0;
+  padding: 1px 7px;
+  justify-content: space-between;
 `;
 
 export default {
@@ -313,12 +177,6 @@ export default {
   Reconnecting,
   VideoContainer,
   Video,
-  Info,
-  Dropdown,
-  UserName,
-  Muted,
-  Voice,
-  DropdownTrigger,
-  PinButtonWrapper,
-  PinButton,
+  TopBar,
+  BottomBar,
 };
