+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/user-status/component.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Styled from './styles';
+
+const UserStatus = (props) => {
+  const { voiceUser } = props;
+
+  const listenOnly = voiceUser?.listenOnly;
+  const muted = voiceUser?.muted;
+  const voiceUserJoined = voiceUser?.joined;
+
+  return (
+    <>
+      {(muted && !listenOnly) && <Styled.Muted iconName="unmute_filled" />}
+      {listenOnly && <Styled.Voice iconName="listen" /> }
+      {(voiceUserJoined && !muted) && <Styled.Voice iconName="unmute" />}
+    </>
+  );
+};
+
+export default UserStatus;
+
+UserStatus.defaultProps = {
+};
+
+UserStatus.propTypes = {
+  voiceUser: PropTypes.shape({
+    listenOnly: PropTypes.bool.isRequired,
+    muted: PropTypes.bool.isRequired,
+    joined: PropTypes.bool,
+  }).isRequired,
+};
