+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/view-actions/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const FullscreenWrapper = styled.div`
+  position: relative;
+`;
+
+export default {
+  FullscreenWrapper,
+};
